i quit
you win css devs
why is css so bad
there is no way to animate DEFAULT height like how can you not animate that? like how do you even not be able to calculate that? what the heck?? this is the reason why everyone uses javascript css is just bad